requirements.txt
pandas
jinja2
weasyprint
python-dotenv
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
streamlit